Move HideAiGeneratedImages to settings-api
