+++ duckchat/duckchat-impl/src/main/java/com/duckduckgo/duckchat/impl/pixel/AiFeaturesStateReporter.kt
@@ -19,7 +19,7 @@ package com.duckduckgo.duckchat.impl.pixel
 import com.duckduckgo.app.statistics.pixels.Pixel
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.duckchat.impl.DuckChatInternal
-import com.duckduckgo.duckchat.impl.store.HideAiGeneratedImages
+import com.duckduckgo.settings.api.HideAiGeneratedImages
 import com.duckduckgo.settings.api.SearchAssistVisibility
 import com.duckduckgo.settings.api.SerpSettingsDataProvider
 import com.squareup.anvil.annotations.ContributesBinding
